change to xmake, refactor vector, add node lib

+++ rasterizer/include/rasterizer.h
@@ -1,8 +1,7 @@
-
 #ifndef RASTERIZER_H
 #define RASTERIZER_H
 #include <graphics.h>
-#include <wmath/wmath.h>
+#include <wmath.h>
 #define WRGB(r, g, b) ((r << 16) | (g << 8) | b)
 
 using namespace WMath;
@@ -10,7 +9,7 @@ using namespace WMath;
 namespace Rasterizer {
     class Rasterizer2D {
         public:
-            static void DrawTriangle(const Vector2d& a, const Vector2d& b, const Vector2d& c, color_t color);
+            static void DrawTriangle(const Vector2i& a, const Vector2i& b, const Vector2i& c, color_t color);
     };
 }
