+++ rasterizer/src/rasterizer.cpp
@@ -1,15 +1,15 @@
-#include <rasterizer/rasterizer.h>
+#include <rasterizer.h>
 
 using namespace Rasterizer;
 using namespace WMath;
 
-void Rasterizer2D::DrawTriangle(const Vector2d& a, const Vector2d& b, const Vector2d& c, color_t color) {
+void Rasterizer2D::DrawTriangle(const Vector2i& a, const Vector2i& b, const Vector2i& c, color_t color) {
 	int top = a.y;
 	int bottom = a.y;
 	int left = a.x;
 	int right = a.x;
 
-	// 计算三角形的包围盒
+	// calculate the bounding box of the triangle
 	if (b.y > top) {
 		top = b.y;
 	}
@@ -38,27 +38,28 @@ void Rasterizer2D::DrawTriangle(const Vector2d& a, const Vector2d& b, const Vect
 		right = c.x;
 	}
 
-	// 使用叉乘进行像素是否在三角形内的判断:
-	// 叉乘的右手定则可以判断出点在直线的左边还是右边，如果三个点都在直线的同一侧，那么叉乘的结果都是正数或者都是负数
-	// 但是叉乘可以进一步优化，因为叉乘的结果在循环中是一个线性变化
-	// 可以通过叉乘的运算方式算得
-	// 总之 cross0 = P0.x * (B.y - A.y) + P0.y * (A.x - B.x) + B.x * A.y - A.x * B.y
+	// use cross product to judge whether a pixel is in the triangle
+	// the right hand rule of cross product can judge whether a point is on the left or right of a line
+	// if all three points are on the same side of the line, the result of cross product is all positive or negative
+	// but cross product can be optimized, because the result of cross product is a linear change in the loop
+	// so we can calculate the cross0 and then add a constant in each loop
+	// cross0 = P0.x * (B.y - A.y) + P0.y * (A.x - B.x) + B.x * A.y - A.x * B.y
 	// crossn = cross0 + (Pn.x - P0.x) * (B.y - A.y) + (Pn.y - P0.y) * (A.x - B.x)
-	// 所以只在开始算一次 cross0，然后每次循环只需要加上一个常数即可
-	/** 已弃用的直接叉乘法
+	// so we only need to calculate cross0 once and then add a constant in each loop
+	/** deprecated code
 	for (int y = bottom; y <= top; y++) {
 		for (int x = left; x <= right; x++) {
-			Vector2d p(x, y);
-			Vector2d v0 = b - a;
-			Vector2d v1 = c - b;
-			Vector2d v2 = a - c;
-			Vector2d v3 = p - a;
-			Vector2d v4 = p - b;
-			Vector2d v5 = p - c;
+			Vector2f p(x, y);
+			Vector2f v0 = b - a;
+			Vector2f v1 = c - b;
+			Vector2f v2 = a - c;
+			Vector2f v3 = p - a;
+			Vector2f v4 = p - b;
+			Vector2f v5 = p - c;
 
-			double cross0 = Vector2d::Cross(v3, v0);
-			double cross1 = Vector2d::Cross(v4, v1);
-			double cross2 = Vector2d::Cross(v5, v2);
+			double cross0 = Vector2f::Cross(v3, v0);
+			double cross1 = Vector2f::Cross(v4, v1);
+			double cross2 = Vector2f::Cross(v5, v2);
 
 			if (cross0 >= 0 && cross1 >= 0 && cross2 >= 0) {
 				putpixel(x, y, color);
@@ -68,7 +69,7 @@ void Rasterizer2D::DrawTriangle(const Vector2d& a, const Vector2d& b, const Vect
 	}
 	**/
 
-	// 优化后的叉乘法
+	// optimized code
 	double l0 = b.y - a.y;
 	double l1 = c.y - b.y;
 	double l2 = a.y - c.y;
